+++ tket/src/OpType/include/OpType/OpType.hpp
@@ -636,6 +636,11 @@ enum class OpType {
   MultiplexedU2Box,
 
   /**
+   * See \ref StatePreparationBox
+   */
+  StatePreparationBox,
+
+  /**
    * See \ref ClassicalExpBox
    */
   ClassicalExpBox,