+++ tket/src/Transformations/include/Transformations/BasicOptimisation.hpp
@@ -24,6 +24,7 @@ namespace Transforms {
 // removes gate-inverse pairs, merges rotations, removes identity rotations,
 // and removes redundant gates before measure Expects: Any gates Produces: The
 // same gate set
+bool redundancy_removal(Circuit& circuit);
 Transform remove_redundancies();
 
 /**