+++ tket/src/Utils/include/Utils/EigenConfig.hpp
@@ -57,6 +57,9 @@ void to_json(nlohmann::json& j, const Matrix<_Scalar, _Rows, _Cols>& matrix) {
 
 template <typename _Scalar, int _Rows, int _Cols>
 void from_json(const nlohmann::json& j, Matrix<_Scalar, _Rows, _Cols>& matrix) {
+  // resize for dynamically-sized matrices
+  if (j.size() == 0) return;
+  matrix.resize(j.size(), j.at(0).size());
   for (size_t i = 0; i < j.size(); ++i) {
     const auto& j_row = j.at(i);
     for (size_t j = 0; j < j_row.size(); ++j) {