+++ pytket/docs/changelog.rst
@@ -1,6 +1,27 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+Major new features:
+
+* New ``StatePreparationBox`` to prepare arbitrary quantum states.
+
+Minor new features:
+
+* New ``CommutableMeasuresPredicate`` predicate, added as precondition to the
+  ``DelayMeasures`` pass.
+* Added an ``allow_partial`` parameter to the ``DelayMeasures`` pass to delay
+  the measurements as much as possible when they cannot be fully delayed to the
+  end.
+* Update to ``pytket-circuit-renderer`` 0.5.
+
+Fixes:
+
+* ``DelayMeasures`` pass now correctly handles circuits with ``CircBox``es.
+
+
 1.11.1 (January 2023)
 ---------------------